pyproject.toml
[tool.poetry]
name = "surge-pricing"
version = "0.1.0"
description = ""
authors = ["Denise Avila <denisefavila@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.2.1"
fastapi = "^0.115.6"
h3 = "^4.1.2"
python-dotenv = "^1.0.1"
dash = "^2.18.2"
plotly = "^5.24.1"
pandas = "^2.2.3"
folium = "^0.19.2"
branca = "^0.8.1"
cassandra-driver = "^3.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
